api configured, image added

+++ src/main/java/com/example/demomaven/controllers/ProductController.java
@@ -1,12 +1,16 @@
 package com.example.demomaven.controllers;
 
 import com.example.demomaven.models.Product;
+import com.example.demomaven.repositories.ProductRepository;
 import com.example.demomaven.services.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
 
 @RestController
@@ -14,31 +18,73 @@ public class ProductController {
 
     @Autowired
     private ProductService productService;
+    @Autowired
+    private ProductRepository productRepository;
 
-    @GetMapping("/products")
-    private ResponseEntity<List<Product>> getAllProducts(){
+    @GetMapping("/product")
+    private ResponseEntity<List<Product>> getAllProducts() {
+
         return new ResponseEntity<>(productService.getAllProducts(), HttpStatus.OK);
     }
 
-    @GetMapping("/products/{id}")
-    private ResponseEntity<Product> getProductById(@PathVariable int id){
-        return new ResponseEntity<>(productService.getProductById(id), HttpStatus.OK);
+    @GetMapping("/product/{id}")
+    private ResponseEntity<Product> getProductById(@PathVariable int id) {
+
+        Product product = productService.getProductById(id);
+        if (product != null) {
+            return new ResponseEntity<>(productService.getProductById(id), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
-    @PostMapping("products")
-    private ResponseEntity<Product> addProduct(@RequestBody Product product){
-        return new ResponseEntity<>(productService.addProduct(product), HttpStatus.OK);
+    @GetMapping("/product/image/{id}")
+    private ResponseEntity<byte[]> getImageByProductId(@PathVariable int id) {
+
+        Product product = productService.getProductById(id);
+        if(product == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        byte[] imageFile = product.getImageData();
+        return ResponseEntity.ok().contentType(MediaType.valueOf(product.getImageType()))
+                .body(imageFile);
     }
 
-    @PutMapping("products")
-    public ResponseEntity<Product> updateProduct(@RequestBody Product product){
-        return new ResponseEntity<>(productService.updateProduct(product), HttpStatus.OK);
+    @PostMapping("product")
+    private ResponseEntity<?> addProduct(@RequestPart Product product,
+                                         @RequestPart MultipartFile imageFile) throws IOException {
+
+        try {
+            Product addedProduct = productService.addProduct(product, imageFile);
+            return new ResponseEntity<>(addedProduct, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+        }
     }
 
-    @DeleteMapping("products/{id}")
-    public void deleteProduct(@PathVariable int id){
-        productService.deleteProduct(id);
+    @PutMapping("product/{id}")
+    public ResponseEntity<?> updateProduct(@PathVariable int id,
+                                                 @RequestPart Product product,
+                                                 @RequestPart MultipartFile imageFile) throws IOException {
+
+            Product updatedProduct = productService.updateProduct(id, product, imageFile);
+            if(updatedProduct != null) {
+                return new ResponseEntity<>(updatedProduct, HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity<>("Failed to update", HttpStatus.BAD_REQUEST);
+            }
     }
 
+    @DeleteMapping("product/{id}")
+    public ResponseEntity<String> deleteProduct(@PathVariable int id) {
 
+        Product product = productService.getProductById(id);
+        if (product != null) {
+            productService.deleteProduct(id);
+            return new ResponseEntity<>("Product is deleted", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Failed to delete", HttpStatus.NOT_FOUND);
+        }
+    }
 }
